Add nvidia-smi util
The nvidia-smi utility is generally bundled with userspace components to
help check hardware detection and features/functionality.  This seems
like a good/reasonable place to add it, exposing via the content
interface to make it usable via that mechanism.

Having the utility available as an app also enables quick
testing/validation one this snap is installed.